+++ src/Udred/Acq/IDataFacade.java
@@ -3,14 +3,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Acq;
+package Udred.Acq;
 
-import Acq.ICase;
-import Udred.Business.Case;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import Acq.IDataFacade;
 
 /**
  *
