+++ src/Udred/Acq/IDiaryEntry.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Acq;
+package Udred.Acq;
 
 import Udred.Business.DiaryEntryTypeEnum;
 import java.util.Date;
