+++ src/Udred/Acq/IGUIFacade.java
@@ -1,0 +1,23 @@
+
+package Udred.Acq;
+
+import java.io.IOException;
+import java.sql.SQLException;
+import javafx.stage.Stage;
+
+/**
+ *
+ * @author JogGez
+ */
+public interface IGUIFacade 
+{
+    
+        void start(Stage primaryStage)throws Exception ;
+        
+        void initialize() throws IOException, SQLException;
+        
+        void injectBusiness(IBusinessFacade bus);
+        
+        ICase getCase(int CaseID);
+
+}
