+++ src/Udred/Acq/IGuardian.java
@@ -3,9 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Acq;
-
-import Udred.Business.Address;
+package Udred.Acq;
 
 /**
  *
