+++ src/Udred/Acq/IMainWindowController.java
@@ -3,10 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Acq;
+package Udred.Acq;
 
-import Acq.IBusinessFacade;
-import Udred.Business.Case;
 import java.io.IOException;
 import java.sql.SQLException;
 import javafx.event.ActionEvent;
