+++ src/Udred/Acq/SYSLogAnnotation.java
@@ -1,4 +1,4 @@
-package Acq;
+package Udred.Acq;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
