+++ src/Udred/Business/CaseInformation.java
@@ -5,7 +5,7 @@
  */
 package Udred.Business;
 
-import Acq.*;
+import Udred.Acq.*;
 import java.util.ArrayList;
 import java.util.List;
