+++ src/Udred/Business/CaseTypeEnum.java
@@ -5,7 +5,7 @@
  */
 package Udred.Business;
 
-import Acq.*;
+import Udred.Acq.*;
 
 /**
  *
