+++ src/Udred/Business/InquiryInformation.java
@@ -5,7 +5,7 @@
  */
 package Udred.Business;
 
-import Acq.*;
+import Udred.Acq.*;
 import java.util.Date;
 
 /**
