+++ src/Udred/Business/PartRepresentive.java
@@ -5,7 +5,7 @@
  */
 package Udred.Business;
 
-import Acq.*;
+import Udred.Acq.*;
 
 /**
  *
