--- src/Udred/Business/Udred.java
@@ -1,25 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Udred.Business;
-
-import java.awt.List;
-import Acq.*;
-
-
-/**
- *
- * @author Linea Hoffmann
- */
-public class Udred
-{
-    private List[] Case;
-    private IDataFacade databaseFacede;
-    
-    public Udred(List Case)
-    {
-        
-    }
-}
